+++ Tella/Data/Database/Common/SqliteError.swift
@@ -8,13 +8,15 @@ import SQLite3
 
 // Indicates an exception during a SQLite Operation.
 class SqliteError : Error {
-    var message = ""
+    var message : String?
     var error = SQLITE_ERROR
     
-    init(message: String = "") {
-        debugLog(message, level: .debug)
+    init(message: String? = nil) {
         self.message = message
+        guard let message else { return }
+        debugLog(message, level: .debug)
     }
+    
     init(error: Int32) {
         self.error = error
     }
