+++ Tella/Scenes/Settings/Views/Servers/AddServer/AddServerURLView.swift
@@ -73,6 +73,13 @@ struct AddServerURLView: View {
         }
         
         .navigationBarHidden(true)
+        
+        .onAppear {
+            
+        #if DEBUG
+                    serverViewModel.projectURL = "https://tella.world/p/dhekra"
+         #endif
+        }
     }
     
     private var serverLoginView: some View {
